Add winget install test back to download cron workflow
This was originally included in the download cron workflow, but had to be removed because there wasn't a reliable way to use the winget CLI from GitHub's runners (it worked locally). There's now a GitHub Action available for installing the winget CLI, which we should be able to use now for this.